% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacenR-package.R
\docType{package}
\name{bacenR-package}
\alias{bacenR}
\alias{bacenR-package}
\title{bacenR: Tools to Access Data from Banco Central do Brasil}
\description{
Provides functions to query, retrieve, and tidy economic and financial data from Banco Central do Brasil web services for use in R analyses and workflows.
}
\author{
\strong{Maintainer}: Ricardo Theodoro \email{rtheodoro@usp.br}

}
\keyword{internal}
