% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacenR-package.R
\docType{package}
\name{bacenR-package}
\alias{bacenR}
\alias{bacenR-package}
\title{bacenR: Tools to Access Data from Banco Central do Brasil}
\description{
Provides functions to query, retrieve, and tidy economic and financial data from Brazilian Central Bank web services for use in R analyses and workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rtheodoro/bacenR}
  \item Report bugs at \url{https://github.com/rtheodoro/bacenR/issues}
}

}
\author{
\strong{Maintainer}: Ricardo Theodoro \email{rtheodoro@usp.br} (\href{https://orcid.org/0000-0001-6741-5153}{ORCID})

}
\keyword{internal}
