% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ifdata_reports.R
\name{get_ifdata_reports}
\alias{get_ifdata_reports}
\title{Download institution reports data from \href{https://olinda.bcb.gov.br/olinda/servico/IFDATA/versao/v1/aplicacao}{IFdata of Brazilian Central Bank (Bacen)} Relatorios}
\usage{
get_ifdata_reports(year, month, type_institution, report, verbose = TRUE)
}
\arguments{
\item{year}{Numeric or vector. year (ex: 2024 or c(2023, 2024))}

\item{month}{Numeric or vector. month (ex: 1 to 12, or c(6, 12))}

\item{type_institution}{Numeric. Type of Institution:
\itemize{
\item 1 = Conglomerados Prudenciais e Instituicoes Independentes
\item  2 = Conglomerados Financeiros e Instituicoes Independentes
\item 3 = Instituicoes Individuais
\item 4 = Instituicoes com Operacoes de Cambio
}}

\item{report}{Numeric. Report type:
\itemize{
\item 1 = Resumo
\item 2 = Ativo
\item 3 = Passivo
\item 4 = Demonstracao de Resultado
\item 5 = Informacoes de Capital
\item 6 = Segmentacao
\item 7 = Carteira de Credito Ativa - Por indexador
\item 8 = Carteira de Credito ativa - por nivel de risco da operacao
\item 9 = Carteira de Credito ativa - por regiao geografica
\item 10 = Carteira de Credito ativa - quantidade de clientes e de operacoes
\item 11 = Carteira de Credito ativa Pessoa Fisica - modalidade e prazo de vencimento
\item 12 = Carteira de Credito ativa Pessoa Juridica - por atividade economica (CNAE)
\item 13 = Carteira de Credito ativa Pessoa Juridica - modalidade e prazo de vencimento
\item 14 = Carteira de Credito ativa Pessoa Juridica - por porte do tomador
\item 15 = Movimentacao de Cambio no Trimestre
\item  16 = Carteira de Credito ativa - por carteiras de instrumentos financeiros
}}

\item{verbose}{Logical. If TRUE, print progress messages (default: TRUE)}
}
\value{
Data frame with IFdata values or NULL in the case of errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Download IFdata reports from the Brazilian Central Bank (Bacen) API for specified years,
months, institution types, and report types. The function handles multiple combinations
of parameters and returns a consolidated data frame with the results.

Also, check the page to see the available reports for each institution type and date: \href{https://www3.bcb.gov.br/ifdata/index2024.html}{IFdata}.
}
\examples{
\donttest{
# Unique institution type for a specific period
data <- get_ifdata_reports(
 year = 2024,
 month = 12,
 report = 1,
 type_institution = 2
)

# Multiple periods, take some time to run
data <- get_ifdata_reports(
  year = c(2023, 2024),
  month = c(6, 12),
  report = 1,
  type_institution = 2
)
}
}
