% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ifdata_reports.R
\name{get_ifdata_reports}
\alias{get_ifdata_reports}
\title{Download institution reports data from \href{https://olinda.bcb.gov.br/olinda/servico/IFDATA/versao/v1/aplicacao}{IFdata of Brazilian Central Bank (Bacen)} Relatorios}
\usage{
get_ifdata_reports(year, month, type_institution, report, verbose = TRUE)
}
\arguments{
\item{year}{Numeric or vector. year (ex: 2024 or c(2023, 2024))}

\item{month}{Numeric or vector. month (1 to 12, or c(6, 12))}

\item{type_institution}{Numeric or vector. Type of Institution:
1 = Conglomerados Prudenciais e Instituicoes Independentes
2 = Conglomerados Financeiros e Instituicoes Independentes
3 = Instituicoes Individuais (default)
4 = Instituicoes com Operacoes de Cambio}

\item{report}{Numeric. Report type:
1 = Resumo
2 = Ativo
3 = Passivo
4 = Demonstracao de Resultado
5 = Informacoes de Capital
6 = Segmentacao
7 = Carteira de Credito Ativa - Por indexador
8 = Carteira de Credito ativa - por nivel de risco da operacao
9 = Carteira de Credito ativa - por regiao geografica
10 = Carteira de Credito ativa - quantidade de clientes e de operacoes
11 = Carteira de Credito ativa Pessoa Fisica - modalidade e prazo de vencimento
12 = Carteira de Credito ativa Pessoa Juridica - por atividade economica (CNAE)
13 = Carteira de Credito ativa Pessoa Juridica - modalidade e prazo de vencimento
14 = Carteira de Credito ativa Pessoa Juridica - por porte do tomador
15 = Movimentacao de Cambio no Trimestre
16 = Carteira de Credito ativa - por carteiras de instrumentos financeiros}

\item{verbose}{Logical. Se TRUE, print progress messages (default: TRUE)}
}
\value{
Data frame with IFdata values or NULL in the case of errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Download IFdata reports from the Brazilian Central Bank (Bacen) API for specified years,
months, institution types, and report types. The function handles multiple combinations
of parameters and returns a consolidated data frame with the results.

Also, check the page to see the available reports for each institution type and date: \href{https://www3.bcb.gov.br/ifdata/index2024.html}{IFdata}.
}
\examples{
\donttest{
# Unique institution type for a specific period
data <- get_ifdata_reports(
 year = 2024,
 month = 12,
 report = 1,
 type_institution = 3
)
}

\dontrun{
# Multiple institution types for a specific period
data <- get_ifdata_reports(
 year = 2024,
 month = 12,
 report = 2,
 type_institution = 4
)

# Multiple institution types for multiple periods
data <- get_ifdata_reports(
  year = c(2023, 2024),
  month = c(6, 12),
  report = 3,
  type_institution = 1
)
}
}
