% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ifdata_values.R
\name{get_ifdata_values}
\alias{get_ifdata_values}
\title{Download data from \href{https://olinda.bcb.gov.br/olinda/servico/IFDATA/versao/v1/aplicacao#!/recursos/IfDataValores#eyJmb3JtdWxhcmlvIjp7IiRmb3JtYXQiOiJ0ZXh0L2NzdiIsIiR0b3AiOjEwMCwiBDIEIjoiMjAxODAzIiwiUmVsYXRvcmlvIjoiVCIsIgQwBCI6IjEyMzQifSwicHJvcHJpZWRhZGVzIjpbMCwxLDIsMyw0LDUsNiw3LDgsOV0sInBlc3F1aXNhZG8iOnRydWUsImFjdGl2ZVRhYiI6InRhYmxlIiwiZ3JpZFN0YXRlIjp7AzADOlt7A0IDIgQwBCIsA0EDfSx7A0IDIgQxBCIsA0EDfSx7A0IDIgQyBCIsA0EDfSx7A0IDIgQzBCIsA0EDfSx7A0IDIgQ0BCIsA0EDfSx7A0IDIgQ1BCIsA0EDfSx7A0IDIgQ2BCIsA0EDfSx7A0IDIgQ3BCIsA0EDfSx7A0IDIgQ4BCIsA0EDfSx7A0IDIgQ5BCIsA0EDfV0sAzEDOnt9LAMyAzpbXSwDMwM6e30sAzQDOnt9LAM1Azp7fX0sInBpdm90T3B0aW9ucyI6ewNhAzp7fSwDYgM6W10sA2MDOjUwMCwDZAM6W10sA2UDOltdLANmAzpbXSwDZwM6ImtleV9hX3RvX3oiLANoAzoia2V5X2FfdG9feiIsA2kDOnt9LANqAzp7fSwDawM6ODUsA2wDOmZhbHNlLANtAzp7fSwDbgM6e30sA28DOiIENgRnZW0iLANwAzoiVGFibGUifX0=}{IF.data of Brazilian Central Bank (Bacen)}}
\usage{
get_ifdata_values(
  year,
  month,
  type_institution = 3,
  relatorio = "T",
  verbose = TRUE
)
}
\arguments{
\item{year}{Numeric or vector. year (ex: 2024 or c(2023, 2024))}

\item{month}{Numeric or vector. month (1 to 12, or c(6, 12))}

\item{type_institution}{Numeric or vector. Type of Institution:
1 = Conglomerados Prudenciais e Instituicoes Independentes
2 = Conglomerados Financeiros e Instituicoes Independentes
3 = Instituicoes Individuais (default)
4 = Instituicoes com Operacoes de Cambio
Or any combination of these (e.g., c(1, 3))}

\item{relatorio}{Character. Type of document: 'T' = Total (default), 'A' = Ativo, 'P' = Passivo}

\item{verbose}{Logical. Se TRUE, print progess messages (default: TRUE)}
}
\value{
Data frame with IF.data values or NULL in the case of errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Downloads IF.data values from the Brazilian Central Bank (Bacen) API for specified years,
months, institution types, and report types. The function handles multiple combinations
of parameters and returns a consolidated data frame with the results.
}
\examples{
\donttest{
# Unique institution type for a specific period
data <- get_ifdata_values(year = 2024, month = 12, type_institution = 3)
}

\dontrun{
# Multiple institution types for a specific period
data <- get_ifdata_values(year = 2024, month = 12, type_institution = c(2, 3))

# Multiple institution types for multiple periods
data <- get_ifdata_values(
  year = c(2023, 2024),
  month = c(6, 12),
  type_institution = c(1, 2, 3)
)
}
}
